Fixing compatibility so that tests pass again on all platforms.

+++ pylsdj/project.py
@@ -20,7 +20,7 @@ def load_lsdsng(filename):
 
     with open(filename, 'rb') as fp:
         # Skip the preamble this time around
-        fp.seek(len(preamble_data) / 8)
+        fp.seek(int(len(preamble_data) / 8))
 
         # Load compressed data into a block map and use BlockReader to
         # decompress it
